Rename in_memory packages and modules to memory.
