+++ src/logicblocks/event/persistence/memory/converters/clause.py
