+++ src/logicblocks/event/persistence/memory/converters/query.py
