+++ src/logicblocks/event/processing/broker/locks/__init__.py
@@ -1,4 +1,4 @@
 from .base import Lock as Lock
 from .base import LockManager as LockManager
-from .in_memory import InMemoryLockManager as InMemoryLockManager
+from .memory import InMemoryLockManager as InMemoryLockManager
 from .postgres import PostgresLockManager as PostgresLockManager
