+++ src/logicblocks/event/processing/broker/subscribers/stores/instances/__init__.py
@@ -1,4 +1,4 @@
 from .base import EventSubscriberStore as EventSubscriberStore
-from .in_memory import (
+from .memory import (
     InMemoryEventSubscriberStore as InMemoryEventSubscriberStore,
 )
