+++ src/logicblocks/event/processing/broker/subscribers/stores/state/__init__.py
@@ -1,6 +1,6 @@
 from .base import EventSubscriberState as EventSubscriberState
 from .base import EventSubscriberStateStore as EventSubscriberStateStore
-from .in_memory import (
+from .memory import (
     InMemoryEventSubscriberStateStore as InMemoryEventSubscriberStateStore,
 )
 from .postgres import (
