+++ src/logicblocks/event/store/adapters/__init__.py
@@ -1,5 +1,5 @@
 from .base import EventSerialisationGuarantee, EventStorageAdapter
-from .in_memory import InMemoryEventStorageAdapter
+from .memory import InMemoryEventStorageAdapter
 from .postgres import PostgresEventStorageAdapter
 from .postgres import QuerySettings as PostgresQuerySettings
