+++ src/logicblocks/event/store/adapters/memory/adapter.py
