+++ src/logicblocks/event/store/adapters/memory/converters.py
