+++ tests/unit/logicblocks/event/projection/store/adapters/test_memory.py
