+++ tests/unit/logicblocks/event/sources/test_memory.py
