+++ tests/unit/logicblocks/event/store/adapters/test_memory.py
